+++ .env.example
@@ -1,0 +1,2 @@
+# Get your API key here: https://resend.com/api-keys
+RESEND_API_KEY=
