+++ .gitignore
@@ -1,3 +1,1 @@
-# Supabase
-.branches
-.temp
+.env
