+++ supabase/.gitignore
@@ -1,0 +1,3 @@
+# Supabase
+.branches
+.temp
